pick an alias that's not a tag
For ARM the linker inserts tags like '$a' and '$a.2' to indicate the beginning
of a chunk of ARM instructions. These appear as aliases of functions in the
output of the `stack_sizes` crate, and sometimes they would appear in the call
graph as the canonical name of the node.

This commit changes how the canonical name is selected to avoid picking these
tags.