get stack usage info from the intermediate object file
rather than from the final executable. This sidesteps the problem of the linker
dropping the stack usage information.

This commit also adds logic to load stack usage information from the
compiler-builtins rlib; this information is not included in the intermediate
object file because this rlib is always linked as a separate object file

Finally, this commit also makes all `cargo-call-stack` invocations rebuild the
crate. Without this it's hard to locate the intermediate object file, which
leads to loading wrong / outdated stack usage information.